Added ComputersTeletype
[This ComputersTeletype extension v9 by Dan Bowen is based on v8 of Computers by Emily Short.
There are only 2 things added: the variables visibleDescriptor and screenDescriptor.
The default to "visible" and "screen".  These appear when a multiple-choice menu is displayed.
In the story you change them to whatever text you like.  In this case, it's "printed" and "paper.]

An extension for Inform 7 to modify the output of computer interactions to appear as a teletype would appear.based on the Computers extension by Emily Short.